Finalize styles; add delete get routes; add contact images

+++ intouch_app/main/routes.py
@@ -1,8 +1,8 @@
 from flask import Blueprint, render_template, redirect, url_for, flash
 from flask_login import login_required, current_user
 from datetime import datetime
-from intouch_app.models import Contact, Comment
-from intouch_app.main.forms import ContactForm, CommentForm
+from intouch_app.models import Contact, Note
+from intouch_app.main.forms import ContactForm, NoteForm, DeleteForm
 
 from intouch_app.extensions import db
 
@@ -29,6 +29,7 @@ def create_contact():
             phone=form.phone.data,
             address=form.address.data,
             birthday=form.birthday.data,
+            image_url = form.image_url.data,
             relationship=form.relationship.data,
             user_id=current_user.id,
         )
@@ -44,7 +45,7 @@ def create_contact():
 @login_required
 def contact_detail(contact_id):
     contact = Contact.query.get(contact_id)
-    comments = Comment.query.filter_by(contact_id=contact_id).all()
+    notes = Note.query.filter_by(contact_id=contact_id).all()
 
     if contact is None or contact.user_id != current_user.id:
         return redirect(url_for("main.homepage"))
@@ -58,6 +59,7 @@ def contact_detail(contact_id):
         contact.phone = form.phone.data
         contact.address = form.address.data
         contact.birthday = form.birthday.data
+        contact.image_url = form.image_url.data
         contact.relationship = form.relationship.data
 
         db.session.commit()
@@ -66,7 +68,7 @@ def contact_detail(contact_id):
         return redirect(url_for("main.contact_detail", contact_id=contact_id))
 
     return render_template(
-        "contact_detail.html", contact=contact, comments=comments, form=form
+        "contact_detail.html", contact=contact, notes=notes, form=form
     )
 
 
@@ -86,6 +88,7 @@ def edit_contact(contact_id):
         contact.phone = form.phone.data
         contact.address = form.address.data
         contact.birthday = form.birthday.data
+        contact.image_url = form.image_url.data
         contact.relationship = form.relationship.data
 
         db.session.commit()
@@ -96,80 +99,86 @@ def edit_contact(contact_id):
     return render_template("edit_contact.html", contact=contact, form=form)
 
 
-@main.route("/contact/<contact_id>/delete", methods=["POST"])
+@main.route("/contact/<contact_id>/delete", methods=["GET", "POST"])
 @login_required
 def delete_contact(contact_id):
     contact = Contact.query.get(contact_id)
     if contact is None or contact.user_id != current_user.id:
         return redirect(url_for("main.homepage"))
 
-    db.session.delete(contact)
-    db.session.commit()
+    form = DeleteForm(obj=contact)
 
-    flash("Contact was deleted successfully.")
-    return redirect(url_for("main.homepage"))
+    if form.validate_on_submit():
+        db.session.delete(contact)
+        db.session.commit()
+
+        flash("Contact was deleted successfully.")
+        return redirect(url_for("main.homepage"))
+    return render_template("delete_contact.html", contact=contact, form=form)
 
 
-@main.route("/contact/<contact_id>/create_comment", methods=["GET", "POST"])
+@main.route("/contact/<contact_id>/create_note", methods=["GET", "POST"])
 @login_required
-def create_comment(contact_id):
+def create_note(contact_id):
     contact = Contact.query.get(contact_id)
     if contact is None or contact.user_id != current_user.id:
         return redirect(url_for("main.homepage"))
 
-    form = CommentForm()
+    form = NoteForm()
 
     # if form was submitted and contained no errors
     if form.validate_on_submit():
-        new_comment = Comment(
-            comment=form.comment.data, created_on=datetime.now(), contact_id=contact.id
+        new_note = Note(
+            note=form.note.data, created_on=datetime.now(), contact_id=contact.id
         )
 
-        db.session.add(new_comment)
+        db.session.add(new_note)
         db.session.commit()
 
-        flash("New comment was created successfully.")
+        flash("New note was created successfully.")
         return redirect(url_for("main.contact_detail", contact_id=contact.id))
 
-    return render_template("create_comment.html", contact=contact, form=form)
+    return render_template("create_note.html", contact=contact, form=form)
 
 
-@main.route("/contact/<contact_id>/comment/<comment_id>/edit", methods=["GET", "POST"])
+@main.route("/contact/<contact_id>/note/<note_id>/edit", methods=["GET", "POST"])
 @login_required
-def edit_comment(contact_id, comment_id):
+def edit_note(contact_id, note_id):
     contact = Contact.query.get(contact_id)
-    comment = Comment.query.get(comment_id)
+    note = Note.query.get(note_id)
     if contact is None or contact.user_id != current_user.id:
         return redirect(url_for("main.homepage"))
 
-    form = CommentForm(obj=comment)
+    form = NoteForm(obj=note)
 
     # if form was submitted and contained no errors
     if form.validate_on_submit():
-        comment.comment = form.comment.data
-        comment.created_on = datetime.now()
+        note.note = form.note.data
+        note.created_on = datetime.now()
 
-        db.session.add(comment)
+        db.session.add(note)
         db.session.commit()
 
-        flash("Comment was updated successfully.")
+        flash("Note was updated successfully.")
         return redirect(url_for("main.contact_detail", contact_id=contact.id))
 
     return render_template(
-        "edit_comment.html", contact=contact, comment=comment, form=form
+        "edit_note.html", contact=contact, note=note, form=form
     )
 
 
-@main.route("/contact/<contact_id>/comment/<comment_id>", methods=["POST"])
+@main.route("/contact/<contact_id>/note/<note_id>", methods=["GET", "POST"])
 @login_required
-def delete_comment(contact_id, comment_id):
+def delete_note(contact_id, note_id):
     contact = Contact.query.get(contact_id)
-    comment = Comment.query.get(comment_id)
+    note = Note.query.get(note_id)
     if contact is None or contact.user_id != current_user.id:
         return redirect(url_for("main.homepage"))
+    form = DeleteForm(obj=contact)
+    if form.validate_on_submit():
+        db.session.delete(note)
+        db.session.commit()
 
-    db.session.delete(comment)
-    db.session.commit()
-
-    flash("Comment was deleted successfully.")
-    return redirect(url_for("main.contact_detail", contact_id=contact.id))
+        flash("Note was deleted successfully.")
+        return redirect(url_for("main.contact_detail", contact_id=contact.id))
+    return render_template("delete_note.html", contact=contact, note=note, form=form)
